Move internal helper classes to new internal support namespace
By placing experimental or internal classes within a dedicated namespace like Hyde\Support\Internal, we make it clear to other developers that these classes are not part of the stable public API. It also allows us to organize our codebase more effectively, making it easier to maintain and evolve over time. It also gives us more freedom to create more helper classes like this, without increasing the maintenance burden.
